Start merging in the candle opsworks commands with deployment commands for opsworks instances.
Renamed the gem to qops so that we don’t need to include a thor file to run it.

Minor corrections to get deployments to work.

Version bump to 0.2.0 plus other updates

Correct issue with names of instances spawned.

Less slack notifications.

Better handling of error cases.

Ensure hostnames are not too long.